chore: 주석 제거 및 네이밍 변경, 파일별 그룹화
