+++ BankManagerUIApp/BankManagerUIApp/BankController/BankViewController.swift
@@ -1,19 +1,15 @@
 //
 //  BankManagerUIApp - ViewController.swift
-//  Created by yagom. 
+//  Created by Eddy, dudu.
 //  Copyright © yagom academy. All rights reserved.
 // 
 
 import UIKit
 
-class ViewController: UIViewController {
+final class BankViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .systemBackground
-        // Do any additional setup after loading the view.
     }
-
-
 }
-
